Fix update image api and adjust portofolio feature (create, update, view

+++ backend/src/feature/portofolio/application/achievement_service.ts
@@ -92,7 +92,8 @@ export class AchievementService {
     if (
       result &&
       FormatHelper.isPresent(data.image_path) &&
-      FormatHelper.isPresent(oldImagePath)
+      FormatHelper.isPresent(oldImagePath) && 
+      data.image_path !== oldImagePath
     ) {
       this.minioService.removeObject(oldImagePath);
     }
