+++ backend/src/feature/portofolio/application/framework_service.ts
@@ -6,9 +6,7 @@ import { MinioService } from '../../support/application/minio_service';
 import { MinioResponse } from '../../support/domain/model/response/minio_response';
 import { FrameworkEntity } from '../domain/model/entities/framework_entity';
 import { CreateFrameworkRequest } from '../domain/model/request/framework/create_framework_request';
-import {
-  FRAMEWORK_DATABASE_REPOSITORY,
-} from '../domain/repository/database/framework_database_repository';
+import { FRAMEWORK_DATABASE_REPOSITORY } from '../domain/repository/database/framework_database_repository';
 import type { FrameworkDatabaseRepository } from '../domain/repository/database/framework_database_repository';
 import { FrameworkRequest } from '../domain/model/request/framework/framework_request';
 import { FrameworkResponse } from '../domain/model/response/framework_response';
@@ -90,7 +88,8 @@ export class FrameworkService {
     if (
       result &&
       FormatHelper.isPresent(data.image_path) &&
-      FormatHelper.isPresent(oldImagePath)
+      FormatHelper.isPresent(oldImagePath) &&
+      data.image_path !== oldImagePath
     ) {
       this.minioService.removeObject(oldImagePath);
     }
